fix(monetary): completar migración Cashier a integer (ADR-018)
PROBLEMA:
Tests de Cashier fallaban porque múltiples entidades y recursos
todavía usaban decimal:2 y (float) casts en lugar de integer.

CAMBIOS APLICADOS:

1. CashCount Entity:
   - 'expected_amount' => 'decimal:2' → 'integer'
   - 'counted_amount' => 'decimal:2' → 'integer'
   - 'difference' => 'decimal:2' → 'integer'
   - Eliminados todos los (float) casts
   - round(..., 2) → (int) round(...)

2. Resources (CashRegisterResource, CashCountResource):
   - Eliminados (float) casts en todos los montos
   - +0 eliminado, reemplazado por (int)

3. Controllers (CashierReportController, CashierDashboardController):
   - Todos los (float) casts en montos → (int)

4. Requests (CreateCashRegisterRequest):
   - 'numeric' → 'integer' para opening_amount_default

5. Tests (CashierEntitiesTest, CashierPayBillTest, CashierServicesTest):
   - Eliminados casts (float) en assertions
   - Expectativas float (95000.0) → int (95000)
   - Expectativas string ('50000.00') → int (50000)

RESULTADO:
✅ Todos los tests de Cashier pasan
✅ Modelo monetario 100% integer en módulo Cashier
✅ Sin casts (float) residuales en código de producción

+++ app/Modules/Cashier/Interfaces/Controllers/CashierDashboardController.php
@@ -74,12 +74,12 @@ class CashierDashboardController extends Controller
         if ($currentSession) {
             $paymentsByMethod = $this->paymentQueryService->getPaymentsByMethodInSession($currentSession->id);
             $totalPayments = $paymentsByMethod->sum('total_amount');
-            $currentBalance = (float) $currentSession->opening_amount + (float) $totalPayments;
+            $currentBalance = (int) $currentSession->opening_amount + (int) $totalPayments;
 
             $sessionData = [
                 'uuid' => $currentSession->uuid,
                 'session_number' => $currentSession->session_number,
-                'opening_amount' => (float) $currentSession->opening_amount,
+                'opening_amount' => (int) $currentSession->opening_amount,
                 'current_balance' => $currentBalance,
             ];
         }
